pyproject.toml
[project]
name = "personal-finance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.124.0",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.1",
    "langchain-google-genai>=3.2.0",
    "matplotlib>=3.10.7",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "seaborn>=0.13.2",
    "uvicorn>=0.38.0",
]
